moved scripts to application directory
