+++ supervision/__init__.py
@@ -15,8 +15,6 @@ from supervision.dataset.core import (
 )
 from supervision.detection.annotate import BoxAnnotator, MaskAnnotator, AnchorAnnotator
 from supervision.detection.core import Detections
-from supervision.detection.line_counter import LineZone, LineZoneAnnotator
-from supervision.detection.tools.polygon_zone import PolygonZone, PolygonZoneAnnotator
 from supervision.detection.utils import (
     box_iou_batch,
     filter_polygons_by_area,
@@ -26,6 +24,8 @@ from supervision.detection.utils import (
     polygon_to_mask,
     polygon_to_xyxy,
 )
+from supervision.application.line_counter import LineZone, LineZoneAnnotator
+from supervision.application.polygon_zone import PolygonZone, PolygonZoneAnnotator
 from supervision.draw.color import Color, ColorPalette
 from supervision.draw.utils import draw_filled_rectangle, draw_polygon, draw_text
 from supervision.geometry.core import Point, Position, Rect
